+++ instrumentation/opentelemetry-instrumentation-asgi/tests/test_asgi_middleware.py
@@ -1093,7 +1093,8 @@ class TestAsgiApplication(AsyncAsgiTestBase):
 
         trace_id = format_trace_id(span.get_span_context().trace_id)
         span_id = format_span_id(span.get_span_context().span_id)
-        traceresponse = f"00-{trace_id}-{span_id}-01"
+        trace_flags = span.get_span_context().trace_flags
+        traceresponse = f"00-{trace_id}-{span_id}-{trace_flags:02x}"
 
         self.assertListEqual(
             response_start["headers"],
@@ -1367,7 +1368,8 @@ class TestAsgiApplication(AsyncAsgiTestBase):
 
         trace_id = format_trace_id(span.get_span_context().trace_id)
         span_id = format_span_id(span.get_span_context().span_id)
-        traceresponse = f"00-{trace_id}-{span_id}-01"
+        trace_flags = span.get_span_context().trace_flags
+        traceresponse = f"00-{trace_id}-{span_id}-{trace_flags:02x}"
 
         self.assertListEqual(
             socket_send["headers"],